eliminate code duplication of logic in process_cvs_changesets()
Factor out code which is specific to rcsparse and cvsclient into
separate functions and pass a parameter to process_cvs_changesets()
so it can decide which of the two needs to be used.

This supersedes the function process_cvs_rlog_changesets() which
duplicated the looping code also contained in process_cvs_changesets().