cvsclient: Allow to connect to a pserver URL without password
The CVS client was raising an error when trying to connect to such pserver
URL: pserver://anonymous@cvs.example.org/cvsroot/project/module

But numerous CVS pserver URLs that can be found in the wild (notably on
SourceForge and OSDN) are in that form.

So add support for such URL form in the CVS client.

Also remove use of external dependency urllib3.util.parse_url and prefer
to use urllib.parse.urlparse from standard Python library.