cvsclient: Align commands to checkout a file with git cvsimport ones
When using pserver protocol, the previous implementation could fail to
checkout some files while it is working fine when using the cvsimport
command from git.

So align the pserver commands sent to checkout a file with the ones used
in the git cvsimport implementation as it effectively fixes the observed
file checkout issues.